// src/schema/gameSystem.d.ts
export type Constraints = Constraint[];
// ...
export type Modifiers = Modifier[];
export type ModifierGroups = ModifierGroup[];
export type Rules = Rule[];
export type CategoryLinks = CategoryLink[];
export type InfoLinks = InfoLink[];
export type Profiles = Profile[];
export type InfoGroups = InfoGroup[];
/**
 * Types of cost this game system uses
 */
export type CostTypes = {
  comment?: string;
  defaultCostLimit: number;
  hidden?: boolean;
  id: string;
  name: string;
}[];
export type Costs = {
  name: string;
  /**
   * References a costType
   */
  typeId: string;
  value: number;
}[];
export type SelectionEntryLinks = SelectionEntryLink[];
export type SelectionEntryGroups = SelectionEntryGroup[];
export type SelectionEntries = SelectionEntry[];
export type ForceEntries1 = ForceEntry[];
export type ForceEntries = ForceEntry[];
/**
 * profile types this game system uses
 */
export type ProfileTypes = {
  comment?: string;
  id: string;
  name: string;
  /**
   * @minItems 1
   */
  characteristicTypes?: [CharacteristicType, ...CharacteristicType[]];
}[];

/**
 * A game system file, data about a system for use with any number of roster building tools
 */
export interface BSDGameSystem {
  /**
   * How to get a hold of the game system's maintaners
   */
  authorContact?: string;
  /**
   * The name(s) of the authors/maintaners
   */
  authorName?: string;
  /**
   * The authors/maintaners website or other URL
   */
  authorUrl?: string;
  /**
   * If this datafile is battlescribe-compatible, which version is required
   */
  battleScribeVersion?: string;
  /**
   * Categories for this game system, automatically imported to every catalogue using it
   */
  categoryEntries?: Category[];
  comment?: string;
  costTypes?: CostTypes;
  /**
   * entryLinks that are available and validated for every force using this game system
   */
  entryLinks?: SelectionEntryLink[];
  forceEntries?: ForceEntries;
  id: string;
  name: string;
  profileTypes: ProfileTypes;
  /**
   * publications that should be listed in every roster using this game system
   */
  publications?: Publication[];
  /**
   * Additional information about this game system
   */
  readme?: string;
  /**
   * Incrementing revision number for this game system
   */
  revision: number;
  rules?: Rule[];
  selectionEntries?: SelectionEntry[];
  /**
   * profiles available to any catalogue using this game system
   */
  sharedProfiles?: Profile[];
  /**
   * rules available to any catalogue using this game system
   */
  sharedRules?: Rule[];
  /**
   * selectionEntries available to any catalogue using this game system
   */
  sharedSelectionEntries?: SelectionEntry[];
  /**
   * infoGroups available to any catalogue using this game system
   */
  sharedInfoGroups?: InfoGroup[];
  /**
   * selectionEntryGroups available to any catalogue using this game system
   */
  sharedSelectionEntryGroups?: SelectionEntryGroup[];
  /**
   * Identifies this object as a gameSystem (as opposed to a catalogue or roster)
   */
  type: "gameSystem";
}
export interface Category {
  comment?: string;
  constraints?: Constraints;
  hidden?: boolean;
  id: string;
  infoLinks?: InfoLinks;
  infoGroups?: InfoGroups;
  modifiers?: Modifiers;
  modifierGroups?: ModifierGroups;
  name: string;
  page?: number | string;
  profiles?: Profiles;
  /**
   * References a publication
   */
  publicationId?: string;
  rules?: Rules;
}
// ...
export interface Rule {
  comment?: string;
  description?: string;
  hidden?: boolean;
  id: string;
  modifiers?: Modifiers;
  modifierGroups?: ModifierGroups;
  name: string;
  page?: number | string;
  /**
   * References a publication
   */
  publicationId?: string;
}
export interface InfoGroup {
  hidden?: boolean;
  id: string;
  infoGroups?: InfoGroups;
  infoLinks?: InfoLinks;
  modifiers?: Modifiers;
  modifierGroups?: ModifierGroups;
  name?: string;
  page?: number | string;
  profiles?: Profiles;
  /**
   * References a publication
   */
  publicationId?: string;
  rules?: Rules;
}
export interface Profile {
  characteristics: {
    "#text"?: string;
    name: string;
    /**
     * References a characteristicType
     */
    typeId: string;
  }[];
  comment?: string;
  hidden?: boolean;
  id: string;
  modifiers?: Modifiers;
  modifierGroups?: ModifierGroups;
  name?: string;
  page?: number | string;
  /**
   * References a publication
   */
  publicationId?: string;
  /**
   * References a profileType
   */
  typeId: string;
  typeName?: string;
}
export interface SelectionEntryLink {
  categoryLinks?: CategoryLinks;
  collective?: boolean;
  comment?: string;
  constraints?: Constraints;
  costs?: Costs;
  entryLinks?: SelectionEntryLinks;
  hidden?: boolean;
  id: string;
  infoLinks?: InfoLinks;
  infoGroups?: InfoGroups;
  modifiers?: Modifiers;
  modifierGroups?: ModifierGroups;
  name?: string;
  page?: number | string;
  profiles?: Profiles;
  /**
   * References a publication
   */
  publicationId?: string;
  rules?: Rules;
  selectionEntries?: SelectionEntries;
  selectionEntryGroups?: SelectionEntryGroups;
  /**
   * References a selectionEntry, selectionEntryGroup or another selectionEntryLink
   */
  targetId: string;
  type?: "selectionEntry" | "selectionEntryGroup";
}
export interface SelectionEntry {
  categoryLinks?: CategoryLinks;
  collective?: boolean;
  comment?: string;
  costs?: Costs;
  constraints?: Constraints;
  entryLinks?: SelectionEntryLinks;
  hidden?: boolean;
  id: string;
  infoLinks?: InfoLinks;
  infoGroups?: InfoGroups;
  modifiers?: Modifiers;
  modifierGroups?: ModifierGroups;
  name: string;
  page?: number | string;
  profiles?: Profiles;
  /**
   * References a publication
   */
  publicationId?: string;
  rules?: Rules;
  selectionEntries?: SelectionEntries;
  selectionEntryGroups?: SelectionEntryGroups;
  type: string;
}
export interface SelectionEntryGroup {
  categoryLinks?: CategoryLinks;
  collective?: boolean;
  comment?: string;
  constraints?: Constraints;
  defaultSelectionEntryId?: string;
  entryLinks?: SelectionEntryLinks;
  hidden?: boolean;
  id: string;
  infoLinks?: InfoLinks;
  infoGroups?: InfoGroups;
  modifiers?: Modifiers;
  modifierGroups?: ModifierGroups;
  name: string;
  page?: number | string;
  profiles?: Profiles;
  /**
   * References a publication
   */
  publicationId?: string;
  rules?: Rules;
  selectionEntries?: SelectionEntries;
  selectionEntryGroups?: SelectionEntryGroups;
}
export interface ForceEntry {
  categoryLinks?: CategoryLinks;
  comment?: string;
  constraints?: Constraints;
  forceEntries?: ForceEntries1;
  hidden?: boolean;
  id: string;
  modifiers?: Modifiers;
  modifierGroups?: ModifierGroups;
  name: string;
  rules?: Rules;
}
export interface CharacteristicType {
  comment?: string;
  id: string;
  name: string;
}
export interface Publication {
  comment?: string;
  id: string;
  name: string;
  publicationDate?: string;
  publisher?: string;
  publisherUrl?: string;
  shortName?: string;
}